continue to work on layout and styles for login and other auth pages
